Move class to each file
